DMP-869 Refactored Search class, moved to pageObjects. Removed commented code.
