+++ src/test/java/uk/gov/hmcts/darts/automation/cucumber/steps/StepDef_search.java
@@ -3,16 +3,16 @@ package uk.gov.hmcts.darts.automation.cucumber.steps;
 import io.cucumber.java.en.Then;
 import uk.gov.hmcts.darts.automation.utils.NavigationShared;
 import uk.gov.hmcts.darts.automation.utils.SharedDriver;
-import uk.gov.hmcts.darts.automation.utils.searchCase;
+import uk.gov.hmcts.darts.automation.pageObjects.Search;
 
 import java.util.List;
 import java.util.Map;
 
 public class StepDef_search extends StepDef_base{
-    private searchCase SRCH;
+    private Search SRCH;
     public StepDef_search(SharedDriver driver, NavigationShared ns) {
         super(driver, ns);
-        SRCH = new searchCase(webDriver);
+        SRCH = new Search(webDriver);
     }
 
     @Then("I can see search results table")
